Better integration with ChromiumOS SDK
This commit only affects the DEKU that is used in the ChromiumOS SDK
environment.

DEKU now requires the kernel must be built with:
USE="livepatch kernel_sources" emerge-<BOARD_NAME>

Eliminate the need to perform synchronization after building the kernel.

DEKU now requires passing the "--board=<BOARD_NAME>" parameter to the
init/build/deploy commands. For example:
./deku --board=brya -p root@localhost init
./deku --board=brya build
./deku --board=brya deploy

The "-d ssh" parameter is not required in the init process.

DEKU can now be used outside of the project directory. This means that
the DEKU directory can be added to the system PATH environment and
performing the "deku …" command can be run anywhere inside the
ChromiumSDK environment. Due to this change, the use of the
“make init/build/sync/deploy” command is now deprecated.

A separate README for ChromiumOS SDK has been added.

+++ command/sync.sh
@@ -15,11 +15,23 @@ regenerateSymbols()
 
 main()
 {
+	local run=$1
+
+	if [[ $run != "auto" && "$KERN_SRC_INSTALL_DIR"  ]]; then
+		logInfo "For this configuration, manual synchronization is not required."
+		return
+	fi
+
 	logInfo "Synchronize..."
 	rm -rf "$workdir"/deku_*
 	getKernelVersion > "$KERNEL_VERSION_FILE"
 	regenerateSymbols
-	git --work-tree="$SOURCE_DIR" --git-dir="$workdir/.git" add "$SOURCE_DIR/*"
+
+	if [ "$KERN_SRC_INSTALL_DIR" ]; then
+		touch -r "$KERN_SRC_INSTALL_DIR" "$KERNEL_VERSION_FILE"
+	else
+		git --work-tree="$SOURCE_DIR" --git-dir="$workdir/.git" add "$SOURCE_DIR/*"
+	fi
 }
 
 main $@
